Add tests for SemanticVersion.ToString and ctor
These tests validate that the constructor throws the expected exceptions on bad
input, and that the return value of [ToString] is as expected. This commit also
adds two methods related to unit testing: AssertThrows(T), for checking for any
exception or its subclasses; and AssertThrowsExact(T) for checking for an exact
exception without counting subclasses.

SemanticVersion now also implements IEquatable(SemanticVersion) and overrides a
few of the virtual methods provided by [object].