Add [SemanticVersion.CompatibleWith] method, with unit tests.
[CompatibleWith] determines whether its parameter is backwards-compatible with the
instance it is called on. There are going to be a few situations where this method
is not suitable (such as comparing 0.x.x versions, as the method always provides
[false] result for these as they are not guaranteed to be compatible).

This commit also changes the [>=] and [<=] operators so that the versions compared
are compared with an equivalency comparison instead of comparing equality. This is
because the operators work with precedence, and exact equality is not prerequisite
for having the same precedence.