Adds a new [SemanticVersion] constructor, signature (int, int).
This constructor is just for convenience, when you have major and minor versions,
but don't have a patch version. This constructor omits the patch version argument
and defaults the value of the [Patch] property to zero.

This commit also changes the constructor unit tests, splitting them into separate
methods: [ConstructingInvalid] for invalid input, and [ConstructingValid] to test
valid input.