Implement [IFormattable] on [SemanticVersion], with tests.
It is conceivable that there may be multiple ways to format a semantic version, and
implementing [IFormattable] can make the job of formatting much more convenient for
the user by exposing a set of pre-defined formats.

This commit includes four formats: default, prefixed default, concise, and prefixed
concise. The default format is the format given in the Semantic Versioning spec and
is equivalent to the old implementation of [ToString]. The concise format omits the
patch version when it is equal to zero, and always omits build metadata. The format
specifiers these formats use are "G" and "C", respectively. The "prefixed" variants
use the lowercase versions of these specifiers, and prefix the output with a letter
"v" (e.g. "v1.5.3" instead of "1.5.3").

To handle the new [ToString] implementations, the [Stringifying] unit test has been
completely rewritten (and, hopefully, much improved).