Add declarations for [SemanticVersion.Parse/TryParse].
Not much in the way of implementation yet, just wanted to get the basic structure
pushed to the repository. The public [Parse] and [TryParse] method bodies are in,
but the internal method that they call is not implemented.

Unit tests for all parts of this will come with the rest of the implementation.