Add implementations for [SemanticVersion.Parse/TryParse], with tests.
These methods expose the Semantic Version parser to users. Each of the two methods
has two overloads: one accepting a [ParseMode] parameter that modifies how parsing
is performed, and another without the parameter which passes [ParseMode.Strict] to
to the other method.

As is convention, [Parse] throws exceptions on failure, and [TryParse] just uses a
bool value instead of throwing. The naïve way to implement this is to have all the
functionality in [Parse] and just have [TryParse] call [Parse] and stop exceptions
[Parse] throws from propagating, but this has a fair amount of overhead (exception
throwing isn't cheap). Instead, both methods internally call a method that returns
an integer/enum status code (which is cheap) and make a decision using that.