Revise [ToString]'s method of concatenating identifiers/metadata.
The old way had two steps: append each identifier followed by a period, and then
remove a character from the end because we would be left with a trailing period.
This method gets rid of the character removal by adding the first identifier and
appending the rest with a prefixing period.