Enable memoization of the results of [SemanticVersion] parsing methods.
The framework was there but unused in the release for v1.0.0, and some testing
revealed that the inclusion of memoization produced a not-insignificant boost.

Tests were performed on an Intel i5-3570K @ 3.4GGHz, and the test involved the
creation of 1000 semantic version strings (1.5.1-beta.0, 1.5.1-beta.1, ...) in
a loop, where each string was passed to [SemanticVersion.Parse] 100,000 times.

The unmemoized test (the code that was released as version 1.0.0) completed in
a time of 589.6 seconds. The memoized test (this commit) took 29.5 seconds, or
560.1s (95%) less time than the unmemoized test. The memoized used more memory
than the unmemoized test, but the extra was only around 400-500kB.

I think an improvement like that justifies the addition of the code.