Add [VersionRange.ParseResult] struct, change access modifiers.
The [ParseResult] struct represents the result of an attempt to parse, and will be
used like the [SemanticVersion.ParseResult] struct. The second main change is that
access modifiers have been switched from 'private' to 'internal' so unit tests may
access and test those items directly instead of via the main public interfaces.