Expand [ParseResultType] + add [CreateException], [GetErrorMessage]
To make the [VersionRange.ParseResult] struct closer in usage to the one used with
[SemanticVersion] (and for general ease of use), the methods [CreateException] and
[GetErrorMessage] are added.

The values for [VersionRange.ParseResultType] have also been expanded so now there
are some actual values in it.

The next step after this is probably the actual implementation of parsing, as this
is laying down the groundwork so it's all (somewhat) nice/easy to use.