Fix and add tests for [VersionRange] parsing.
The previously-committed code for parsing was untested, and didn't work properly.
This commit includes fixes to this code and unit tests to ensure correct function
in future.

There are also some minor changes to make the code nicer, such as making the main
parse loop a [foreach] loop instead of a [for] loop.