Implement [Parse/TryParse], add constructors, begin testing comparisons.
Constructors for [VersionRange] have been added to allow construction. Presently, the
only public constructor takes a string argument that is the version range string. The
[Parse] method calls this constructor, while [TryParse] does its own thing because it
doesn't want to throw exceptions.

The first unit test for version ranges is also in. It relies on another of the things
added in this commit, the 'Ranges\VersionRange.Comparison.cs' file, which is where we
put all of the implementations of comparison functions, as well as the stuff that one
can use to exchange [ComparisonToken]s from the parser for [IComparator]s that can be
used to do actual comparing.