Revise locking as introduced by 7f2795c.
It is my understanding that 'lock (...)' is syntactic sugar over the
[Monitor.Enter] and [Exit] methods. Thus, with the null coalescing
operator in use, it may be possible that, between the [Enter] call and
the [Exit] call, [MemoizationAgent] is assigned to and the result of
the coalescion changes.

This may or may not be the case, and I expect that explicitly adding a
temporary variable used only for the object on which to lock will be
optimised out by the compiler if it isn't.

A cursory glance over the IL generated in release mode shows that there
might be a significant difference in code, so I'm going to go for the
option which appears safer to me.