Replace usage of [System.Numerics.BigInteger], plus minor changes.
o [Internals] namespace is added with [PseudoBigInt] class, which makes
  makes comparisons based on the numeric values strings represent but
  without providing any "real" numeric/arithmetic functionality.

o [System.Numerics] reference is removed and uses of the [BigInteger]
  structure from that library are replaced with [PseudoBigInt].

o Tests for [PseudoBigInt] are added to the testing assembly.

---

o Make use of the new [out]-related C# 7 feature in the SV formatter.

o Make the testing assembly target .NET 4.6.2, and import from NuGet
  the [System.ValueTuple] package to make some tests easier to work with.