Target .NET Standard and Core, close issue #2.
This commit squashes and merges in support for .NET Standard
and Core, and use of the new [.csproj] format. Provided is a
summary of changes.

o Convert main and test assemblies to use the new format for
  [.csproj] files. Main assembly targets .NET Frameworks 4.5
  and 4.6, .NET Standard 1.0, and .NET Core 1.0; the testing
  assembly targets are as above except .NET Standard.

o READMEs are adjusted to reflect new targets, and dates for
  copyright notices are in places adjusted to show 2017.

o Shims are added to allow .NET Standard/Core builds to work
  as they provide only a subset of the full Framework. These
  shims aren't user-visible, so the changes aren't breaking.

o The tests for [BinaryFormatter] are now only compiled when
  building for the .NET Framework, as it isn't included with
  .NET Standard/Core.

o Minor formatting/[using] statement clean-ups are made.