Make [VersionRange] implement [IComparator].
The interface [VersionRange.IComparator] is used to provide a single
interface for comparators. By having [VersionRange] implement it any
advanced range syntax features can be decomposed to a [VersionRange]
rather than implementing dedicated comparators for it.

If it turns out that isn't a particularly good way to to do it, this
change can easily be reverted.