Fix bug highlighted by 47a56f2.
The rules for 'node-semver' comparisons differ slightly from the
typical rules for Semantic Version comparisons. If a version has
pre-release identifiers, it cannot satisfy a version range other
than one with the same major-minor-patch trio and pre-release
identifers. For example, '1.2.3-alpha.5' > '1.2.2' using normal
comparison rules, but not using version range rules.

The [VersionRange] class failed to adhere to this rule, and would
perform a normal-rules comparison if the range had identifiers but
the comparison operand did not.

This commit corrects this behaviour by introducing an exclusive-or
test for the presence of identifiers. It also makes a minor change
so that the [VersionRange] greater-than comparison is implemented
using the less-than comparison, which avoids duplication.