Rewrite [VersionRange] parser.
This is part of the work towards issue #11.

The revised parser, I think, will be better suited to modification to
support the 'node-semver' advanced range syntax. The previous version
looked to me as if modifying it to support infix operators would be a
painful experience, and I had code quality concerns (in particular,
the necessary use of [goto]s in the version string parsing states).

It probably the most significant bit of rework I see being needed. It
will need to be followed up, I believe, with:

 o Further parser rework to produce [IComparator]s as its output and
   not [ComparatorToken]s. This will support the decomposition of any
   advanced-syntax ranges to [VersionRange] instances.

 o Consequential changes to the supporting code.

 o Rework to enable the parser to work with infix operators.

The first two points I don't see as especially significant. They will
most likely involve moving calls to [VersionRange.Comparator.Create]
into the parser, and removing those calls from outside the parser.