Rework [VersionRange.Parser] to output comparators.
This is part of the changes necessary to enable the parser to support
the advanced version range syntax. It makes the translation of parser
data to [VersionRange]s a part of the parser's supporting code rather
than part of the code consuming the parser's output.

The parser right now will always call through to [UnaryComparator] to
source an implementation. When binary infix operators are introduced,
it will be possible to determine whether to source a unary comparator
or a binary comparator using the state of the parser.