Further fix bug highlighted by 47a56f2.
The correction applied in 581d44e was based on a misreading of the
specification for 'node-semver' comparisons:

 o That specification does not allow a comparison between a version
   and a comparator set unless one of the comparators in the set has
   a major-minor-patch trio which matches and pre-release identifiers

 o This was misinterpreted as requiring all comparators in a set to
   meet those conditions to be satisfied

This commit should properly implement comparisons. Logic is moved
from the comparison implementations and into the [SatisfiedBy] method
of [VersionRange], which enables checking the proper 'any' condition
instead of the incorrect 'all' condition.

Minor changes to tests are necessary to relocate tests which were,
under the misreading, configured to expect the wrong result.