Rework caret operator, etc. implementation scheme.
While ad47247 added a valid implementation of the [VersionRange]
caret operator, the way it was implemented meant that each comparison
created a new [SemanticVersion] instance. It was considered that this
could prove expensive if many comparisons are done.

Reworked in this commit, the [UnaryComparator] class now internally
uses a factory to create advanced-syntax comparators. As delegates
can capture variables, no new [SemanticVersion] need be created on
each call to the comparator. This means the cost is incurred in the
parser rather than in the comparison, which is preferable.