Add tests for [VersionRange] tilde operator.
Unlike the basic operators and the caret/dash operators, the tilde is
an operator the behaviour of which changes depending on how a version
it operates on is written. That is, '~1' is not the same as '~1.0',
although '~1.0' and '~1.0.0' are equivalent. That seems like a stupid
decision, but I've committed to supporting 'node-semver'.

The downside of supporting this operator is that it will require some
changes to the parser to allow a [VersionRange] to determine how the
version string was written.

The [VersionRange] class needs to know when an optional component of
the version is missing. My current thinking is to define extra values
for [ParseMode] and have the parser set trio components negative. As
these normally cannot be negative, it allows [VersionRange] to easily
identify that a component was omitted.

To avoid unintended behaviour with [ParseMode.Lenient], these extra
values might be presented in a multi-bit field around the high-order
bits of a [ParseMode], which is represented by an [int]. If each of
the values requires that the sign bit is cleared, then [Lenient] will
not interfere. Its definition as [~0] means it will set the sign bit.