Add internal [ParseMode] values (and tests).
As b0ab0c3 discussed, my thinking on supporting the extra features
that advanced-syntax [VersionRange] operators require was to have
a set of flags only exposed internally passed with [ParseMode] to
the [SemanticVersion] parser.

This commit adds definitions for the internal modes and tests for
their functionality, as follows:

 o The [OptionalMinor] mode mirrors [ParseMode.OptionalPatch], and
   allows the minor version component to be omitted ('1' == '1.0.0')

 o The [IndicateOmits] causes the parser to set any component that
   was omitted to a negative value so that [VersionRange] can detect
   whether or not the component was omitted

There is a slight issue in that versions with negative components
will leak to users if they use [MemoizationAgent], but I don't think
this will be a great problem. Anything memoizing parse results isn't
going to be processing them.