Implement [OptionalMinor], [IndicateOmits] modes.
The real change in this commit is that the [SemanticVersion] parser
has been rewritten. I took some time to mull over how best to add the
modes to the parser, and I couldn't see a way of doing it with the
old parser without a lot of repetition. Rewritten, the parser is now
organised into reusable functions and should be easier to read.

To allow [IndicateOmits] to work, a new [SemanticVersion] constructor
has been added. While this doesn't feel like the best way to achieve
what was needed, it does allow private fields to retain [readonly]
modifiers, which I felt was important.

This commit also includes corrections to tests for the new modes.