Add prelim. [VersionRange] hyphen operator impl.
This implements the bulk of what's needed to support the hyphen
operator. Crucially, however, it misses out two things:

 o The implementation of the comparison itself

 o Support in the parser for left-hand versions with omitted minor
   and patch versions

As we can't know whether we're dealing with a binary comparator until
after we've already parsed the first version, we can't apply flags to
allow the omission of version components and indicate when they were
omitted. This will require a bit of a reshuffle from what's in this
commit, so I'm keeping a record of what I did using this commit.