Implement [VersionRange] hyphen operator.
This commit makes the necessary changes following on from 0e3f585.

The problem highlighted in that commit is solved by deferring parsing
until it can be known whether or not a binary comparator is being
parsed. This meant removing an early invalidity check and repeating
it wherever needed in the parser, which isn't ideal as it means it'll
be easier to make mistakes in future. However, at least if a solution
comes to me it can be changed at any time.

This commit also makes minor corrections to tests.