Add [ParseMetadata] member to [SemanticVersion].
I wasn't satisfied with the solution of making components negative to
indicate that they were omitted. It felt a bit hackish and, if
nothing else, it doesn't transfer well to supporting X-ranges with
wildcards instead of omissions. The [SemanticVersion] class now has a
member for providing parse metadata. This won't impact the function
of [MemoizationAgent], and should maintain binary compatibility. Not
being publicly exposed, it should also aid development in future.

This change also means the [IndicateOmits] flag is no longer needed.

This commit makes other minor modifications in consequence, as well
as some minor housekeeping changes.