Implement [VersionRange] wildcard operator, etc.
The wildcard operator, called 'X-ranges' by 'node-semver', allows any
version with changes at the same level or a level subordinate to the
wildcard level. It might be used, for example, to filter for only
patch-level changes.

This commit also adds a further set of tests for the wildcard and the
hyphen operators, and makes necessary corrections as well as other
minor changes.