Remove [ParseMode.Greedy] from [Lenient].
On reflection, including [Greedy] in [Lenient] doesn't seem like the
behaviour that would be expected. To me, at least, the plain meaning
of "lenient" would be that the rules still exist (and should broadly
be followed), but that some flexibility exists in how you might get
there. This doesn't line up with the behaviour of [Greedy].

In contrast, [Greedy] throws out the rules. The cases where input is
considered an error are few, and the mode actively ignores anything
that might cause an error unless it can't produce anything sensible
from what it has. In my view, that sort of thing shouldn't be bundled
with a general "let some things slide" flag, as it has the potential
to cause confusion with a difficult-to-identify source.

As such, this commit excludes [Greedy] from [Lenient].