Implement greedy parsing.
This is (generally) a relatively simple change. Where the parser has
a [return] for an error which is fine in greedy parsing, the value of
the [ParseResultType] is passed through a method that checks to see
whether greedy parsing is enabled. If it is, it discards its argument
and returns [ParseResultType.Success]. Otherwise, it relays its arg.

This commit also makes a number of vaguely related (or not) changes:

 o The [Helper] methods for checking metadata item/pre-release
   identifier validity are rewritten to avoid using a regex. Not
   sure why I originally decided on that for such a simple check

 o Renames the 'USE_SHIMS' constant to 'COMMON_SHIMS'. Originally I
   intended to add some LINQ shims specific to .NET Standard (which
   would require a separate constant). However, I decided on another
   method so didn't need the LINQ shims. I think this name is more
   appropriate, though.