Add [Parse]/[TryParse] overload as #12 considered.
The advantage of a greedy parse mode is that it allows you to have
some information after the [SemanticVersion]. That advantage isn't
very useful if you don't know where the valid [SemanticVersion] ends,
as then you can't locate it without implementing a parser of your
own. To avoid that situation, it would be helpful for parser's caller
to be able to take over wherever the [SemanticVersion] parser stops.

This commit adds two overloads, one of [Parse] and one of [TryParse],
that expose the [IEnumerator<char>] used internally by the parser in
reading input. This enables a caller to pick up immediately after the
parser to do whatever the caller needs.

The downside of [IEnumerator<T>] is that, given an instance, it isn't
possible to tell whether the current character is valid. While this
might have been solved by returning [IEnumerable<T>], that would be
significantly more complex and probably involve a fair bit of wheel
reinventing. To mitigate this issue, the returned enumerator is null
if the parser reached the end of the string. This seemed like a good
compromise, and almost makes intuitive sense.