Add [SemanticVersion] formatter benchmarks.
This commit adds very basic benchmarks for the formatter, testing the
fast path for the general format specifier against the slow path with
the formatter. The results locally are:

 o Fast path: 789.3ns, 736 bytes allocated
 o Slow path: 955.8ns, 496 bytes allocated

That's about a 48% increase in allocations for a 21% increase in the
speed of formatting. The numbers are fairly small, so I'm inclined to
leave it as-is at this point.