Mark internal parse info field optional.
This might have broken cross-version deserialisation before, which is
not good. Marking it with [OptionalField] means that, if it isn't
present in an old-serialised instance, it won't result in a a newer
library version throwing an exception on deserialisation.

Also adds an [OptionalField] shim for .NET Standard 1.0, as it
doesn't include the required namespace.