Restructure [VersionRange] for [CompareTo] impl.
The way [VersionRange] comparators were structured before, not enough
information was retained for [CompareTo] to work without effectively
repeating the implementations of the comparators. Broadly, changes in
this commit restructure [VersionRange] to hold necessary information.

More specifically, this commit:

 o Adds [SemanticVersion.MaxValue] and [MinValue] fields, with their
   documentation.

 o Adds [IComparable<SemanticVersion>] to the [IComparator] internal
   interface, allowing comparators to provide their own methods that
   the higher-level [VersionRange.CompareTo] can defer to.

 o Adds a [ContiguousComparator] class from which [UnaryComparator]
   and [BinaryComparator] derive, providing a common implementation
   where almost every comparator (except 'less than' and 'greater
   than') is decomposed into an inclusive range 'm <= x <= n'.

 o For 'less than' and 'greater than' comparators, adds a special
   class [ComparatorShell], which takes delegates to implement the
   [IComparator] interface in these two special cases.

The effect of this on performance will probably be minimal. The worst
contributor will be the introduction of an abstract class, but apart
from this roughly the same number of objects are created as before.

This commit knowingly breaks the following tests:

 o [RangeIntlParsingTests.BasicIdentification]

 o [RangeIntlParsingTests.MultipleComparatorParsing]

 o [RangeIntlParsingTests.MultipleSetParsing]

These tests rely on inspecting hidden implementation detail in the
instances of [IComparator] returned, using it to directly verify the
output of the parser (rather than relying on functionality tests). It
may be reasonable to delete these tests as functionality tests should
indirectly verify the same behaviour.