Add preliminary [CompareTo] implementation.
This implementation passes many tests (156 of 234). Of those that it
doesn't pass, all but 3 appear related to issues #19 and #20.

Two of those three are surprising. They are:

 o '0.7.2-beta' > '0.7.x'

 o '1.3.0-alpha' < '>1.2.3'

These are tests taken directly from 'node-semver', and at the moment
they don't make sense to me. To start with, I would have said that it
isn't possible for both to be true at once.

Following on, '0.7.x' ---> '>=0.7.0 <0.8.0', and so '0.7.2-beta' does
not seem like it should be greater (because it is lower in precedence
than '0.8.0'). The code in this commit produces a result that says as
much, which is the result I would expect.

For the second case, the Semantic Versioning spec says precedence is
based on the first difference evaluating two versions left-to-right
which, if you ask me, is the minor version. Although '1.3.0-alpha' is
a pre-release version, its minor version is above '1.2.3' and so it
should have higher precedence. As '>1.2.3' includes anything with a
precedence higher than '1.2.3', versions with a higher precedence
than '1.3.0-alpha' will be matched and so I would expect comparison
to say 'neither greater nor lesser', i.e. 0. This is the result that
this code gives, but the 'node-semver' tests say otherwise.

It might be worth raising an issue with 'node-semver'.